requirements.txt
langchain==1.2.7
langchain-core==1.2.7
langchain-community==1.2.7
langchain-groq==0.1.4
langchain-huggingface==1.2.0
langchain-astradb==1.0.0
datasets
pypdf
python-dotenv
pandas
flask
prometheus_client 
